requirements.txt
streamlit==1.35.0
google-generativeai==0.5.4
python-dotenv>=1.0.0
